orchestrator/requirements.txt
motor
beanie
loguru
PyYAML
kafka-python
aiokafka
grpcio
grpcio-tools
asyncio